Set表的Score的float指定精度

+++ Hangfire.MySql/Install.sql
@@ -120,7 +120,7 @@ CREATE TABLE `<tableprefix>_Set` (
   `Id` int(11) NOT NULL AUTO_INCREMENT,
   `Key` varchar(100) NOT NULL,
   `Value` varchar(256) NOT NULL,
-  `Score` float NOT NULL,
+  `Score` float(12, 0)) NOT NULL,
   `ExpireAt` datetime DEFAULT NULL,
   PRIMARY KEY (`Id`),
   UNIQUE KEY `IX_Set_Key_Value` (`Key`,`Value`)
